+++ Sveti/Support/Reusable UIComponents/Onboarding/Models/Interfaces/IOnboardingController.swift
@@ -4,4 +4,6 @@ protocol IOnboardingController {
   var viewModel: IOnboardingVM { get set }
   func presentIfNeeded(from parent: UIViewController)
   func move(to: OnboardingMoveDirection)
+  func trackEarlyExit()
+  func trackOnboardingFinished()
 }
